# contracts/openapi.json
{
  "components": {
    "schemas": {
...
      "EmptyTaskCommand": {
        "additionalProperties": false,
        "properties": {},
        "title": "EmptyTaskCommand",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "NavigationTaskStatus": {
        "enum": [
          "open",
          "assigned",
          "in_progress",
          "completed",
          "cancelled"
        ],
        "title": "NavigationTaskStatus",
        "type": "string"
      },
...
      "OutcomeCommandCreate": {
        "properties": {
          "disposition": {
            "$ref": "#/components/schemas/OutcomeDisposition"
          },
          "note": {
            "anyOf": [
              {
                "maxLength": 4000,
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          }
        },
        "required": [
          "disposition"
        ],
        "title": "OutcomeCommandCreate",
        "type": "object"
      },
      "OutcomeCommandRead": {
        "properties": {
          "cancelled_task_ids": {
            "items": {
              "format": "uuid",
              "type": "string"
            },
            "title": "Cancelled Task Ids",
            "type": "array"
          },
          "disposition": {
            "$ref": "#/components/schemas/OutcomeDisposition"
          },
          "need_id": {
            "format": "uuid",
            "title": "Need Id",
            "type": "string"
          },
          "note": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Note"
          },
          "outcome_id": {
            "format": "uuid",
            "title": "Outcome Id",
            "type": "string"
          },
          "recorded_at": {
            "format": "date-time",
            "title": "Recorded At",
            "type": "string"
          },
          "recorded_by_user_id": {
            "format": "uuid",
            "title": "Recorded By User Id",
            "type": "string"
          }
        },
        "required": [
          "outcome_id",
          "need_id",
          "disposition",
          "note",
          "recorded_by_user_id",
          "recorded_at",
          "cancelled_task_ids"
        ],
        "title": "OutcomeCommandRead",
        "type": "object"
      },
      "OutcomeDisposition": {
        "enum": [
          "resolved",
          "closed_unresolved"
        ],
        "title": "OutcomeDisposition",
        "type": "string"
      },
      "OutcomePreviewRead": {
        "properties": {
          "need_id": {
            "format": "uuid",
            "title": "Need Id",
            "type": "string"
          },
          "tasks": {
            "items": {
              "$ref": "#/components/schemas/OutcomePreviewTaskRead"
            },
            "title": "Tasks",
            "type": "array"
          }
        },
        "required": [
          "need_id",
          "tasks"
        ],
        "title": "OutcomePreviewRead",
        "type": "object"
      },
      "OutcomePreviewTaskRead": {
        "properties": {
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "title": {
            "title": "Title",
            "type": "string"
          }
        },
        "required": [
          "id",
          "title",
          "status"
        ],
        "title": "OutcomePreviewTaskRead",
        "type": "object"
      },
...
      "TaskClaimCreate": {
        "additionalProperties": false,
        "properties": {
          "due_at": {
            "format": "date-time",
            "title": "Due At",
            "type": "string"
          },
          "proposed_change_id": {
            "format": "uuid",
            "title": "Proposed Change Id",
            "type": "string"
          }
        },
        "required": [
          "proposed_change_id",
          "due_at"
        ],
        "title": "TaskClaimCreate",
        "type": "object"
      },
...
      "TaskCommandErrorDetail": {
        "properties": {
          "code": {
            "enum": [
              "need_closed",
              "task_state_conflict",
              "task_claim_mismatch",
              "task_unbound",
              "proposal_not_approved",
              "concurrent_change"
            ],
            "title": "Code",
            "type": "string"
          },
          "message": {
            "title": "Message",
            "type": "string"
          }
        },
        "required": [
          "code",
          "message"
        ],
        "title": "TaskCommandErrorDetail",
        "type": "object"
      },
      "TaskCommandErrorResponse": {
        "properties": {
          "detail": {
            "$ref": "#/components/schemas/TaskCommandErrorDetail"
          }
        },
        "required": [
          "detail"
        ],
        "title": "TaskCommandErrorResponse",
        "type": "object"
      },
      "TaskCommandRead": {
        "properties": {
          "assignee_user_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Assignee User Id"
          },
          "authorized_proposed_change_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Authorized Proposed Change Id"
          },
          "completed_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Completed At"
          },
          "due_at": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Due At"
          },
          "follow_up_request_id": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Follow Up Request Id"
          },
          "need_id": {
            "format": "uuid",
            "title": "Need Id",
            "type": "string"
          },
          "replayed": {
            "title": "Replayed",
            "type": "boolean"
          },
          "status": {
            "$ref": "#/components/schemas/NavigationTaskStatus"
          },
          "task_id": {
            "format": "uuid",
            "title": "Task Id",
            "type": "string"
          }
        },
        "required": [
          "task_id",
          "need_id",
          "status",
          "assignee_user_id",
          "due_at",
          "authorized_proposed_change_id",
          "completed_at",
          "follow_up_request_id",
          "replayed"
        ],
        "title": "TaskCommandRead",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "Synthetic demonstration API contract. It is not for patient care and does not accept real PHI.",
    "title": "Oncology Journey Command Center API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/navigator/needs/{need_id}/outcome-preview": {
      "get": {
        "operationId": "get_outcome_preview_v1_navigator_needs__need_id__outcome_preview_get",
        "parameters": [
          {
            "in": "path",
            "name": "need_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Need Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutcomePreviewRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Outcome Preview",
        "tags": [
          "navigator"
        ]
      }
    },
    "/v1/navigator/needs/{need_id}/outcomes": {
      "post": {
        "operationId": "post_outcome_v1_navigator_needs__need_id__outcomes_post",
        "parameters": [
          {
            "in": "path",
            "name": "need_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Need Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": {
              "maxLength": 255,
              "minLength": 1,
              "title": "Idempotency-Key",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutcomeCommandCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutcomeCommandRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Post Outcome",
        "tags": [
          "navigator"
        ]
      }
    },
...
    "/v1/navigator/tasks/{task_id}/claim": {
      "post": {
        "operationId": "claim_task_v1_navigator_tasks__task_id__claim_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TaskClaimCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCommandRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCommandErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Claim Task",
        "tags": [
          "navigator"
        ]
      }
    },
    "/v1/navigator/tasks/{task_id}/complete": {
      "post": {
        "operationId": "complete_task_v1_navigator_tasks__task_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyTaskCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCommandRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCommandErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Task",
        "tags": [
          "navigator"
        ]
      }
    },
    "/v1/navigator/tasks/{task_id}/start": {
      "post": {
        "operationId": "start_task_v1_navigator_tasks__task_id__start_post",
        "parameters": [
          {
            "in": "path",
            "name": "task_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Task Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EmptyTaskCommand"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCommandRead"
                }
              }
            },
            "description": "Successful Response"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskCommandErrorResponse"
                }
              }
            },
            "description": "Conflict"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Task",
        "tags": [
          "navigator"
        ]
      }
    },
...
  }
}